Initial installer script for standalone linux hosts (#45)
* Initial installer script for collector and fluentd

* Initial config with hostmetrics for linux hosts

* Allow release to test stage without signing

* Add default apt/yum repo definition files

* Migrate tests to pytest

+++ internal/buildscripts/packaging/fpm/preuninstall.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
-if command -v systemctl >/dev/null 2>&1; then
+if command -v systemctl >/dev/null 2>&1 && [ -f /usr/lib/systemd/system/splunk-otel-collector.service ]; then
+    echo "Stopping splunk-otel-collector service"
     systemctl stop splunk-otel-collector.service
     systemctl disable splunk-otel-collector.service
 fi
